Prompt for confirmation before deleting anything
Not pretty by any stretch, but it'll suffice for now. Eventually this
can be done up with a proper dimmed 'lightbox' style modal prompt.

+++ resources/js/components/Files/Browser/FileRow.vue
@@ -63,6 +63,9 @@ export default {
             this.$router.push(route);
         },
         remove(file) {
+            if (!confirm("Deletion is permanent! Are you sure?")) {
+                return;
+            }
             this.$store.dispatch('files/delete', file);
         },
     },
